util: Use stat() on files under /sys/fs/resctrl/info
We are skipping non-directories under /sys/fs/resctrl/info/ since those are not
interesting for us.  However in tests it can sometimes happen that ent->d_type
is 0 instead of 4 (DT_DIR) for directories.

I've seen it fail on two machines.  Different machines, different systems, I
cannot reproduce it even using the same setup.  So one of the ways how to work
around this is call stat() on it.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
Message-Id: <024a75c710d412d1a32e42782b69239f9c7c6d30.1517240003.git.mkletzan@redhat.com>