tests: Remove caps from DO_TEST_PARSE_ERROR() calls
The lack of certain capabilities is going to cause errors pretty
much exclusively when building the QEMU command line, because we
purposefully leave such checks to the very last moments before
starting a guest; so it doesn't make much sense to pass QEMU
capabilities to DO_TEST_PARSE_ERROR() call which are supposed,
as the name implies, to result in much earlier error.

The only exception is the "cpu-hotplug-granularity" test, due
to the fact that the checks we want to fail are skipped entirely
if the relevant QEMU feature is not available.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20180130181006.30568-1-abologna@redhat.com>