qemu: Convert jobInfo stats into a union
Convert the stats field in _qemuDomainJobInfo to be a union. This
will allow for the collection of various different types of stats
in the same field. While doing this, also change the name of the
field from @stats to @migStats to make it easier to find.

When starting the async job that will end up being used for stats,
set the @statsType value appropriately. The @mirrorStats are
special and are used with @migStats in order to generate the
returned job stats for a migration.

Using the NONE should avoid the possibility that some random
async job would try to return stats for migration even though
a migration is not in progress.

For now a migration and a save job will use the same statsType

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20180201232443.13923-3-jferlan@redhat.com>