qemu: Introduce qemuProcessHandleDumpCompleted
Handle a DUMP_COMPLETED event processing the status, stats, and
error string. Use the @status in order to copy the error that
was generated whilst processing the @stats data. If an error was
provided by QEMU, then use that instead.

If there's no async job, we can just ignore the data.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20180201232443.13923-7-jferlan@redhat.com>
Reviewed-by: Jiri Denemark <jdenemar@redhat.com>