Fix release of locks with dlm plugin
The dlm lock plugin couldn't release locks when the vm instance
was destroyed, kill manually because that we can't use file
descriptor which couldn't adopt after libvirtd reboot to be
aware of the process dead, we need to proactively to release
the lock just like unlink the <domain>.xml and <domain>.pid
files. This operation could be found in libxl.

Signed-off-by: river <lfu@suse.com>
Message-Id: <1517821676-14399-3-git-send-email-lfu@suse.com>