conf: Check for user aliases duplicates only
When validating a device XML config we check if user provided
alias is unique. We do this by maintaining a hash table of device
aliases as we iterated over all devices defined for the domain.
However, it may happen that what appears as two devices in domain
XML is in fact just one interface in hypervisor. For instance in
qemu driver this is true for uhci/ehci controllers. In that case
an error is reported even though it is not actually an error. At
any rate, we can assume libvirt generated aliases to be unique
and thus really check user provided ones only.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <1c72961bd02f0f93ecd6a5476bd6694e2fab2b67.1517928018.git.mprivozn@redhat.com>