qemu: Alter condition to avoid possible NULL deref
Commit 'f0f2a5ec2' neglected to adjust the if condition to split
out the possibility that the @watchdog is NULL when altering the
message to add detail about the model.

Just split out the condition and use previous/original message, but
with the new message code.

Found by Coverity

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20180209153338.22031-1-jferlan@redhat.com>
Reviewed-by: Chen Hanxiao <chenhanxiao@gmail.com>