qemu: stop passing virConnectPtr into qemuMonitorStartCPUs
There is a long standing hack to pass a virConnectPtr into the
qemuMonitorStartCPUs method, so that when the text monitor prompts
for a disk password, we can lookup virSecretPtr objects. This causes
us to have to pass a virConnectPtr around through countless methods
up the call chain....except some places don't have any virConnectPtr
available so have always just passed NULL. We can finally fix this
disasterous design by using virGetConnectSecret() to open a connection
to the secret driver at time of use.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20180209172419.3089-3-berrange@redhat.com>
Reviewed-by: John Ferlan <jferlan@redhat.com>