Turn on -Wswitch-enum and fix all problems it identifies
As a general rule any time we switch() on something that is an enum, we
want to have a case for every enum constant. The -Wswitch warning will
report any switch where we've violated this rule, except if that switch
has a default case.

Unfortunately it is reasonable to want to list all enum constants *and*
also have a default case. To get a warning in that scenario requires
that we turn on -Wswitch-enum. When we do this, we then have to then
fixup various switches that used an enum but relied on a default: to
silence warnings. Mostly this involved adding missing enum cases, but
in some places it was more practical to silence the warning by casting
the enum to an int, since we don't care about covering all enums.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20180213165707.29888-2-berrange@redhat.com>