qemu: Generate alias and socket path for pr-helper
While we're not generating the command line just yet (look for
the next commits), we can generate the alias for pr-manager.
A domain can have up to one managed pr-manager (in which case
socket path is decided by libvirt and pr-helper is spawned by
libvirt too), but up to ndisks of unmanaged ones (one per disk
basically). In case of the former we can generate a simple alias
and be sure it'll not conflict. But in case of the latter to
avoid any conflicts let's generate alias that's based on
something unique - like disk target.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <bedfb77c87fb83c044ccce0036cf2aa79f9c4974.1519235041.git.mprivozn@redhat.com>