docs: logos: Resize the 'Drop shadow' filter bounding box
Inkscape makes "Drop shadow" filter's bounding box unnecessarily large.
Since inkscape will make a union of all the objects to form an image,
which means that our logos end up having an undesired "space" padding
from one side only which in turn makes usage of vertical alignment of
objects clumsy, since it the spacing would always make it look as out of
alignment. The official way to "fix" this is to fiddle with the
coordinates of the box and adjust its dimensions, not ideal as it lacks
precision, but makes a difference.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <763da99313168c6bc475509d256d870bd7065424.1519598619.git.eskultet@redhat.com>