libvirt-domain: add new virDomainGetSevVmMeasurement() API
The virDomainGetSevVmMeasurement() can be used to retrieve the measurement
of encrypted VM launched using AMD SEV feature. The measurement is a
signature of the memory contents that can be sent to the guest owner as
an attestation that the memory was encrypted correctly by the firmware
before booting the guest.

Signed-off-by: Xiaogang Chen <Xiaogang.Chen@amd.com>
Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>
Message-Id: <20180226175336.79815-5-brijesh.singh@amd.com>