qemu: Call virDomainDefCompatibleDevice when live add cdrom/floppy
Prior to calling qemuDomainChangeEjectableMedia when adding
the cdrom/floppy to the guest, make a second pass at the
virDomainDefCompatibleDevice with the old_disk that's being
updated by the call.

This is similar to the qemuDomainChangeDiskLive path from
qemuDomainUpdateDeviceLive prior to making the same call to
qemuDomainChangeEjectableMedia when the cdrom/floppy is
updated rather than added.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20180301130339.7158-3-jferlan@redhat.com>